[#MHV-36371] refactor message list view
