+++ src/applications/mhv/secure-messaging/containers/FolderListView.jsx
@@ -1,0 +1,87 @@
+import { VaSearchInput } from '@department-of-veterans-affairs/component-library/dist/react-bindings';
+import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { getMessages } from '../actions/messages';
+import useInterval from '../hooks/use-interval';
+import InboxListView from '../components/MessageList/InboxListView';
+import MessageFolderHeader from '../components/MessageList/MessageFolderHeader';
+
+const FolderListView = () => {
+  const dispatch = useDispatch();
+  const error = null;
+  const messages = useSelector(state => state.sm.messages?.messageList);
+  const folder = useSelector(state => state.sm.folders.folder);
+
+  useInterval(() => {
+    dispatch(getMessages(folder.folderId));
+  }, 5000);
+
+  let content;
+  if (messages === undefined) {
+    content = (
+      <va-loading-indicator
+        message="Loading your secure messages..."
+        setFocus
+      />
+    );
+  } else if (messages.length === 0) {
+    // this is a temporary content. There is a separate story to handle empty folder messaging
+    content = (
+      <va-alert status="error" visible>
+        <h2 slot="headline">No messages</h2>
+        <p>There are no messages in this folder</p>
+      </va-alert>
+    );
+  } else if (error) {
+    content = (
+      <va-alert status="error" visible>
+        <h2 slot="headline">We’re sorry. Something went wrong on our end</h2>
+        <p>
+          You can’t view your secure messages because something went wrong on
+          our end. Please check back soon.
+        </p>
+      </va-alert>
+    );
+  } else {
+    content = (
+      <>
+        <InboxListView messages={messages} />
+      </>
+    );
+  }
+
+  return (
+    <div className="vads-l-grid-container">
+      <div className="main-content">
+        <MessageFolderHeader folder={folder} />
+        {/* <h1>Messages</h1>
+        <p className="va-introtext vads-u-margin-top--0">
+          When you send a message to your care team, it can take up to 3
+          business days to get a response.
+        </p>
+        <EmergencyNote />
+        <p className="vads-u-margin-top--0p5 vads-u-margin-bottom--0">
+          <a
+            className="vads-c-action-link--blue compose-message-link"
+            href="/my-health/secure-messages/compose"
+          >
+            Compose message
+          </a>
+        </p> */}
+        <div className="search-messages-input">
+          <label
+            className="vads-u-margin-top--2p5"
+            htmlFor="search-message-folder-input"
+          >
+            Search the Messages folder
+          </label>
+          <VaSearchInput label="search-message-folder-input" />
+        </div>
+
+        <div>{content}</div>
+      </div>
+    </div>
+  );
+};
+
+export default FolderListView;
